Users Table * Adding new user + Editing exist user + Deleting exist user * some fixes

+++ FightClub/GameData/Interfaces/IRepository.cs
@@ -16,10 +16,10 @@ namespace GameData.Interfaces
     public interface IRepository<T> where T : class
     {
         IEnumerable<T> GetAll();
-        T GetById(int id);
+        T GetById(Int64 id);
         IEnumerable<T> Find(Func<T, Boolean> predicate);
         void Create(T item);
         void Update(T item);
-        void Delete(int id);
+        void Delete(Int64 id);
     }
 }
