+++ FightClub/GameData/Repositories/BattleLogRepository.cs
@@ -32,7 +32,7 @@ namespace GameData.Repositories
             return db.BattleLogs;
         }
 
-        public BattleLog GetById(int id)
+        public BattleLog GetById(Int64 id)
         {
             return db.BattleLogs.Find(id);
         }
@@ -52,7 +52,7 @@ namespace GameData.Repositories
             db.Entry(item).State = EntityState.Modified;
         }
 
-        public void Delete(int id)
+        public void Delete(Int64 id)
         {
             BattleLog log = db.BattleLogs.Find(id);
             if (log != null)
